# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Technical Test JARIVIS - IQBAL IKHLASUL AMAL",
        "contact": {
            "name": "Iqbal Ikhlasul Amal",
            "email": "iqbaliamal@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/kompetensi-dasar-detail/list": {
            "get": {
                "tags": [
                    "Kompetensi Dsar Detail"
                ],
                "summary": "Retrieve all kompetensi-dasar-details",
                "operationId": "16a70a42f94da19cd57acc92c3791c74",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kompetensi-dasar-detail/store": {
            "post": {
                "tags": [
                    "Kompetensi Dsar Detail"
                ],
                "summary": "Create a new kompetensi-dasar-detail",
                "operationId": "ff73910b2ff63a1f4c6c6e0a2efdc93f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kompetensi_dasar_id",
                                    "name"
                                ],
                                "properties": {
                                    "kompetensi_dasar_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "kompetensi_dasar_id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        },
                                        "deleted_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kompetensi-dasar-detail/detail/{id}": {
            "get": {
                "tags": [
                    "Kompetensi Dsar Detail"
                ],
                "summary": "Retrieve a kompetensi-dasar-detail",
                "operationId": "9b50f2aa6b974fc3992399ae0bfb31dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "kompetensi-dasar-detail ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kompetensi-dasar-detail/update/{id}": {
            "put": {
                "tags": [
                    "Kompetensi Dsar Detail"
                ],
                "summary": "Update a kompetensi-dasar-detail",
                "operationId": "0f411cfde2b05250e6c3b4fa89131d2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "kompetensi-dasar-detail ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kompetensi_dasar_id",
                                    "name"
                                ],
                                "properties": {
                                    "kompetensi_dasar_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "kompetensi_dasar_id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        },
                                        "deleted_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kompetensi-dasar-detail/delete/{id}": {
            "delete": {
                "tags": [
                    "Kompetensi Dsar Detail"
                ],
                "summary": "Delete a kompetensi-dasar-detail",
                "operationId": "866e9bd8bd81345d208fb51ce9c7c56c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "kompetensi-dasar-detail ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}